Auto-increment patch names
Rather than naming all pull requests `prose-patch`, name it `prose-patch-N` where `N` is the lowest possible number not already in use.

This generates the patch name in `forkRepo` and must also retrieve the patch name (from the refs passed to the callback) in `sendPatch`.

See #257 for additional discussion on the problem.